requirements.txt
streamlit
google-generativeai
pandas
sqlite3
python-dotenv